Fix deprecation warnings and related bugs
* Includes fix from #9 and fixes build
  * Add missing .hasAttr() interface method to test class
* Unsuppress fixable deprecation warnings and fix all warnings
  * Use CorsConfigBuilder instead of CorsConfig.Builder
  * Use HttpHeaderNames instead of HttpHeaders.Names
  * Use .x() instead of .getX() methods of HttpRequest
  * Use .isClient() instead of comparing with JdkSslClientContext.class
  * Use new EmbeddedChannel() instead of null for test

The most notable changes are the use of the lowercase versions of HTTP
header field names (HttpHeaderNames instead of HttpHeaders.Names),
because these are case-insensitive, but HTTP/2 converts them to
lowercase as part of the protocol. Any code that looks at headers and
doesn't treat them in a case-insensitive way (like putting them in maps
without using the normalized lowercase form) was already wrong in
HTTP/1.x and prone to bugs (and probably will result in an error with
HTTP/2) and should be fixed in subsequent PRs. These become more obvious
when the code uses the mixed-case versions in HttpHeaders.Name instead
of the all-lowercase names in HttpHeaderNames and tests break, but they
can also break in production if the HTTP client sends them in an
unexpected case for HTTP/1.x and qonduit code doesn't handle it
properly.